+++ client-app/src/desktop/tabs/home/roadmap/RoadmapGroupItem.js
