+++ client-app/src/desktop/tabs/home/roadmap/RoadmapView.scss
