+++ client-app/src/desktop/tabs/home/roadmap/RoadmapViewItem.js
